changed xmlvalidator to an interface, so further types of concrete xml validation classes can be added later
