+++ src/test/java/com/ch/conversion/validation/XmlValidatorImplTest.java
@@ -12,7 +12,7 @@ import org.mockito.Mockito;
 /**
  * Created by elliott.jenkins on 21/04/2016.
  */
-public class XmlValidatorTest extends TestHelper {
+public class XmlValidatorImplTest extends TestHelper {
     ITransformConfig config;
 
     @Before
@@ -23,8 +23,8 @@ public class XmlValidatorTest extends TestHelper {
     @Test(expected = XmlException.class)
     public void invalidXml() throws Exception {
         String valid = getStringFromFile(INVALID_FORM_XML_PATH);
-        XmlValidator validator = new XmlValidator(config, valid);
-        validator.validate();
+        XmlValidatorImpl validator = new XmlValidatorImpl();
+        validator.validate(config,valid);
     }
 
     @Test(expected = XsdValidationException.class)
@@ -33,36 +33,36 @@ public class XmlValidatorTest extends TestHelper {
         ITransformConfig schemaConfig = Mockito.spy(config);
         Mockito.when(schemaConfig.getSchemasLocation()).thenReturn("testschemas");
 
-        String valid = getStringFromFile(VALID_DS01_XML_PATH);
-        XmlValidator validator = new XmlValidator(schemaConfig, valid);
-        validator.validate();
+        String invalid = getStringFromFile(INVALID_DS01_XML_PATH);
+        XmlValidatorImpl validator = new XmlValidatorImpl();
+        validator.validate(config,invalid);
     }
 
     @Test(expected = XmlException.class)
     public void noFormTypeInXml() throws Exception {
         String valid = getStringFromFile(BLANK_FORM_XML_PATH);
-        XmlValidator validator = new XmlValidator(config, valid);
-        validator.validate();
+        XmlValidatorImpl validator = new XmlValidatorImpl();
+        validator.validate(config,valid);
     }
 
     @Test(expected = XsdValidationException.class)
     public void noXsdForFormType() throws Exception {
         String valid = getStringFromFile(NO_SCHEMA_XML_PATH);
-        XmlValidator validator = new XmlValidator(config, valid);
-        validator.validate();
+        XmlValidatorImpl validator = new XmlValidatorImpl();
+        validator.validate(config,valid);
     }
 
     @Test
     public void validXmlAgainstXsd() throws Exception {
         String valid = getStringFromFile(VALID_DS01_XML_PATH);
-        XmlValidator validator = new XmlValidator(config, valid);
-        validator.validate();
+        XmlValidatorImpl validator = new XmlValidatorImpl();
+        validator.validate(config,valid);
     }
 
     @Test(expected = XsdValidationException.class)
     public void invalidXmlAgainstXsd() throws Exception {
         String invalid = getStringFromFile(INVALID_DS01_XML_PATH);
-        XmlValidator validator = new XmlValidator(config, invalid);
-        validator.validate();
+        XmlValidatorImpl validator = new XmlValidatorImpl();
+        validator.validate(config,invalid);
     }
 }
